Include Heka dashboard plugin on all nodes
For some visibility about how Heka is performing. This will be particularly
useful when we first deploy Heka and learn how it works at scale. We might
decide to disable it at a later date. As far as I can tell it doesn't impact
the performance of Heka itself.

The plugin class binds it to loopback by default and we're not creating any
UFW rules, so this won't be visible outside of the machine. You can SSH port
forward with `ssh -L 4352:localhost:4352 <machine>` to view it.