Fix dependencies of graphite patching
We patch graphite by downloading a file. However the place we download
it to only properly exists if the Govuk::Mount['/opt/graphite'] resource
has been realised. Unfortunately this then creates problems downloading
the file on a Vagrant VM as we skip all the Govuk::Mount resources.

By downloading the file to /tmp instead we have the ability to install
graphite in Vagrant.