Heka output to Graphite over UDP
Temporarily change Heka's `CarbonOutput` plugin to use UDP instead of TCP.

This is used by every machine that is parsing Nginx logs. I discovered that
if Carbon is down (the VM was shutdown in my case) then the output plugin
fails to write and doesn't release/recycle packs. This causes back pressure
on the filter/encoder/input that's feeding it and causes Heka to stop
reading new log entries, meaning that they don't get written to ES/Kibana
either.

By switching this to UDP the writes become async and won't block the
pipeline. It does however mean that we'll throw away metrics if Graphite is
down or there's a network partition. I think this is an acceptable trade-off
at the moment, because log entries are more important.

This doesn't happen for the `TCPOutput` that we use to ship log entries
between machines because it buffers to disk when it can't write. They're
planning to add the same feature for all output plugins in the near future,
but it's not supported right now. I gathered all of this info from:

- https://mail.mozilla.org/pipermail/heka/2014-October/000215.html
- https://mail.mozilla.org/pipermail/heka/2014-October/000218.html