Raise nofile limit for www-data user for graphite
We are no longer running graphite daemons as root, so they will be
subject to a more restrictive ulimit for the number of open files.

The current setting (1024) leads to occasional errors in the
`/var/log/upstart/carbon-aggregator.log`:

```
Could not accept new connection (EMFILE)
```

The fix for this (as described in the warning in
http://graphite.readthedocs.org/en/latest/carbon-daemons.html) is to
raise the number of allowed file descriptors.