Fix conflicting Heka logstreamer differentiators
The `differentiator` options for error log files must not be the same as
access log files because the `LogstreamerInput` plugin uses this name for
the journal file. If they overlap then it will keep rewriting the journal
file with alternating entries, causing Heka to forget where it was with one
of the files.