Install 'camptocamp/augeas' module
Augeas is a pre-requisite for the elasticsearch/elasticsearch Puppet module.
For details about Augeas, have a look at
http://en.wikipedia.org/wiki/Augeas_%28software%29 and http://augeas.net/

This commit adds the module to the Puppetfile, and includes it in the
Elasticsearch node definition so that we're only ever including it when we're
applying Puppet to the elasticsearch-{1,2,3} machines. We don't include it
in the 'govuk_elasticsearch' module because those customisations are applied
to other Elasticsearch clusters, such as the logs Elasticsearch one.