Add Augeas Gem path to Ruby loadpath
In order to ensure that subsequent Puppet runs after the initial run
introducing Augeas complete successfully, without a type/provider error from
Puppet because the Augeas provider couldn't be loaded, we should add the
Augeas Gem's path to the LoadPath.

Adding as a separate commit because it was a bit of a head-scratcher. Also:
commit-message kudos to @dancarley.