Increase PostgreSQL setting wal_keep_segments to 128
For the last 3 days, the transition-postgresql slave in Staging hasn't
been able to recover after the data sync, with lots of these errors:

    FATAL:  could not receive data from WAL stream: ERROR:  requested
        WAL segment 000000010000003800000065 has already been removed

This indicates that a WAL segment has been recycled on the master
before it could be applied to the slave, and so replication can never
catch up. The database is left in an inconsistent state and Bouncer
can't connect. The only way to fix the slave is to resync it, and we
don't want to have to do that every day.

wal_keep_segments is the minimum number of segments to keep on the
master before recycling them. They're 16MB each so 128 is about 2GB;
we currently have over 50GB free on all our postgres masters.

A more permanent fix for this issue would be to set up WAL archiving
to somewhere accessible from the slave, so that segments can be
applied by replication without them having to be kept on the master.