Set Elasticsearch version in repo_version
In order for the Elasticsearch module to be able to manage the installation
of Elasticsearch, we need to ensure that we set a value in the
`repo_version` variable. Not setting a version means that the module will
refuse to manage the repository.

Providing an undef default because we don't want to set this if the module
isn't managing the installation of Elasticsearch itself because we'll only
ever expose versions we want to install in the Apt repositories we run.