Add collecd plugin for etcd
Add a collectd plugin for etcd that uses the cURL-JSON plugin to query
etcd's API.

See this page for more information on the metrics queried:
https://coreos.com/docs/distributed-configuration/etcd-api/

With our current configuration, this will produce Graphite metrics such
as:

    etcd-3_management_development.curl_json-etcd_self.recvBandwidthRate

Some of the metrics queried are scoped to the cluster and so will be
duplicated for each etcd node that collectd runs on; other metrics are
local to each node. etcd exposes both types of metrics on the same
endpoints, so you'll need to check the documentation (above) to see
which is which.

Alternatively, we could change the plugin or type
instance to group metrics in a more intuitive way but the advantage of
the grouping implemented here is that it mirrors etcd's API.

Note that we don't query the `/v2/stats/leader` endpoint because it only
returns statistics for follower nodes (not the leader node), and if we
query the node local to collectd, etcd will respond with a HTTP
`Location` header redirecting us to the leader, which the collectd
cURL-JSON plugin does not support.

The `/v2/stats/leader` endpoint provides some useful metrics such as
failed RPC requests and node latency, so it might be worth trying to add
this in future.