Upgrade puppetlabs/stdlib to 4.5.0
puppetlabs/stdlib 4.2.0 contained a fix to allow merge() to work with empty
strings. This is required because the default value of the config class
param in the elasticsearch/elasticsearch module is undef. This is then
parsed as an empty string by Puppet during catalogue compilation.

Since 4.5.0 is the latest version, we should upgrade straight to that. There
appear (according, at least, to testing and a read through stdlib's
changelog) to be no breaking changes between 4.2.0 and 4.5.0.