Reduce the varnish cache size
There's about 500Mb of space used by the router, statsd, etc.  The cache
recently ran out of space and restarted, so lets reduce the size
allocation by a 500Mb constant to try and avoid this.

For vagrant environments, this could result in a negative allocation of
space, so for these environments just set a size of 100Mb.