Disable access logging for Heka ES proxy
Allow the access logs to be turned off for the Elasticsearch reverse proxy
that goes through Nginx. Errors will still be logged.

Enable this for Heka because it's output is incredibly high volume and each
log entry results in another message for Heka to consume. The *technically*
correct solution might be to stop Heka picking up these logs, but it's not
easier to exclude them from the logstreamer regex and it will still result
in many log entries on disk.