Copy PostgreSQL backups from master server
Each PostgreSQL server (masters and slaves) take a local backup using
the `autopostgresqlbackup` script.

Taking backups on a PostgreSQL slave can cause query conflicts as
described in the PostgreSQL [documentation][], which we are currently
affected by. In particular, backups taken from the slave are often
interrupted leading to incomplete backups.

These backups will be copied to the backup-1.management machine and
consequently the off-site backup machines.

Because we already have backups on the master servers, use those backups
rather than the ones from slaves. This is an interim measure to ensure
we have working backups. See this Pivotal story for more information:

https://www.pivotaltracker.com/story/show/86297242

[documentation]: http://www.postgresql.org/docs/current/static/hot-standby.html#HOT-STANDBY-CONFLICT