Send an accept header with check_app_up command
The healthcheck path for email-alert-api only responds to JSON requests.
The Nagios check for whether an app is running makes a request to the
healthcheck path without any accept headers, and gets a 406 in response
which causes the check to fail.

This commit adds an "Accept: */*" header to the command so that the
application will serve a JSON response instead of the 406.