Stop logstream resources for Nginx access logs
This is similar to f0f9601857d13d60cd71d0db4302e81689196b0b for error logs.
I missed this in the original PR #2412 :(

It has also highlighted something else that I've missed, which is that we
don't want Heka to pick up logs for signon. I'll had to address that
separately.