[Bouncer] MHRA reverse proxy initial configuration
This isn't complete yet but gives us a good place to test from
and gets something into production. In particular there are some
PDF document groups which we may need to add later.

The non-www domain also isn't included here: currently mhra.gov.uk/
redirects to www.mhra.gov.uk/, but everything else on the bare domain
404s, so no one should be expecting to see proxied content at these
paths on the bare domain.

This set of path prefixes have come from MHRA and our investigation
of their site structure. They are moving the content which is staying
behind to new shorter paths - we are proxying only the new locations
so that we can see clearly what isn't yet working there. They will map
the old locations to redirect to the new ones.

The nginx config is based on the UKBA reverse proxy which was in
place for a few months last year; for comparison, it was removed in
https://github.gds/gds/puppet/pull/1733.

We may want to consider increasing the timeouts (some parts of the
old site are slow) but we'll see how it works in production first.