Assign a virtual IP address to frontend-lb nodes
Assign a virtual ('floating') IP address to the frontend-lb nodes with a
view to replacing the vShield Edge load balancers for HTTP/HTTPS
requests.

The virtual IP will be managed by keepalived and will float between
nodes in the `frontend-lb` class. One node, the master, will advertise
the virtual IP address and all other nodes will act as backup nodes.
Only one machine, the master, will receive traffic on this IP address at
one time.

The virtual router ID must be an integer between 1 and 255. We checked
for multicast traffic in our Preview and Staging environments and did
not find any, so we don't think that there are any other virtual router
advertisements on those networks.

Adding this virtual IP address will not affect our internal request
routing until a corresponding change is made to our hosts file, which we
will do in a separate commit once we have tested the virtual IP added
here.

We're not too concerned about adding the password for the virtual router
here at this point; we will review that decision once we've tested that
this works.