Allow VRRP multicast traffic through UFW
Tell UFW to allow VRRP multicast traffic, whose destination is
`244.0.0.18`, the IANA-assigned multicast address for VRRP [1].

We wanted to specify a protocol of `vrrp` to further limit this firewall
rule, but UFW does not seem to support it. This is in contrast to
iptables, which will support any protocol listed in `/etc/protocols`.

[1]: https://tools.ietf.org/html/rfc5798#section-5.1.1.2