Prevent www.wip.mhra.gov.uk DNS record being cached indefinitely
The DNS record for www.wip.mhra.gov.uk is switched between two public IPs every
20 seconds:
* 83.98.30.74
* 83.98.30.20

Each of these IP addresses relates to a specific backend server.

On a weekly basis, MHRA republish their content. This process breaks the
ability for the server being updated to handle requests. Therefore, they change
the DNS record to stay on the other IP address whilst republishing.

With the mechanism we were using to proxy requests, nginx was resolving the
host at startup and then caching the IP until reloaded or restarted. This is
bad because it could mean that we are proxying back to an IP which can't handle
requests. Additionally, they should be free to change the IP addresses of their
hosting without us needing to do anything.

I tested this in development by running puppet and editing the hosts file on my
Mac to point www.mhra.gov.uk at my VM. I could see that for a given location
block, nginx was varying the IP address it used to pass the requests back to,
after around 20 seconds (the TTL for the DNS record). This implies it was
resolving the domain name on the fly. Hopefully it was caching it for the length
of the TTL.

Based upon:
http://forum.nginx.org/read.php?2,215830,215832#msg-215832