Remove unnecessary stubbed facts for Elasticsearch
The third-party `elasticsearch` module which `govuk_elasticsearch` calls
needs the `Linux` fact to be stubbed so that `elasticsearch::params` can
determine the appropriate user/group for that platform.

The other facts listed in here look like the names of parameters on the
`elasticsearch` class rather than the names of facts (they aren't generic
enough). The tests don't fail without them, so I'm removing them for
clarity.