Move logs Redis river to govuk_elasticsearch
Creating a Redis river for logs-elasticsearch is specific to GOV.UK,
so move the river class from elasticsearch_old to govuk_elasticsearch.

As a result, move the estools package installation into the same
module.

In places that we depend on estools (like creating a river) we've made
those dependencies explicit.