Increase metric coverage for PostgreSQL
We don't currently collect metrics for the sum of all connections to
databases on a PostgreSQL server, nor do we collect metrics specific to
the `postgres` database. This commit adds both.

Note that the `<database>` block is repeated twice; once for global
metrics and a second time for queries specific to the `postgres`
database, which pulls in all of the default queries that the Collectd
plugin provides [1].

I'm hoping that this will help us to troubleshoot these errors:

    DB "transition_production" FATAL: remaining connection slots are reserved for non-replication superuser connections

Tested in Vagrant.

[1]: https://collectd.org/documentation/manpages/collectd.conf.5.shtml#backends