Remove govuk_apt::use_uk_ubuntu_mirrror
I'm about to make a change that will enable Puppet to manage
`/etc/apt/sources.list`, which can replace the functionality provided by
this class.

Also, I'm not sure why we would do this exclusively on the development
VM as all of our servers are UK-based, so it makes sense for them all to
use the UK servers.