Manage `/etc/apt/sources.list` with Puppet
Purge `/etc/apt/sources.list` and manage the Apt sources it used to
provide using Puppet. The main aim of this commit is to complement
gds/govuk-provisioning@2ea8f7a, which removed the use of Apt mirrors in
our vCloud Launcher preamble.

The Apt sources defined here will create the following files:

    matt➜~/govuk/vagrant-govuk(master✗)» vagrant ssh graphite-1.management -c "cat /etc/apt/sources.list.d/ubuntu*"
    #file generated by puppet
    # ubuntu-backports
    deb http://gb.archive.ubuntu.com/ubuntu/ precise-backports main restricted universe multiverse
    #file generated by puppet
    # ubuntu
    deb http://gb.archive.ubuntu.com/ubuntu/ precise main restricted universe multiverse
    #file generated by puppet
    # ubuntu-security
    deb http://gb.archive.ubuntu.com/ubuntu/ precise-security main restricted universe multiverse
    #file generated by puppet
    # ubuntu-updates
    deb http://gb.archive.ubuntu.com/ubuntu/ precise-updates main restricted universe multiverse
    Connection to 127.0.0.1 closed.
    matt➜~/govuk/vagrant-govuk(master✗)»

For comparison, this is the `/etc/apt/sources.list` on the
`jumpbox-2.management.preview` box:

    mattbostock@preview-jumpbox-2:~$ cat /etc/apt/sources.list
    deb mirror://mirrors.ubuntu.com/mirrors.txt trusty main restricted universe multiverse
    deb mirror://mirrors.ubuntu.com/mirrors.txt trusty-updates main restricted universe multiverse
    deb mirror://mirrors.ubuntu.com/mirrors.txt trusty-backports main restricted universe multiverse
    deb mirror://mirrors.ubuntu.com/mirrors.txt trusty-security main restricted universe multiverse
    mattbostock@preview-jumpbox-2:~$

I've defaulted to the UK mirrors as all our servers are UK-based.

Also note, we don't use a list of Apt mirrors because this has caused us
problems in the past due to unreliable mirrors. See
gds/govuk-provisioning@2ea8f7a for more details.