[#87097350] Passive check for automysqlbackup
Add a Nagios passive check for `automysqlbackup` which performs local
backups of MySQL data.

We have existing checks (in `backup::directory`) to confirm whether backups
are copied from individual MySQL servers to central `backup-1.management`
but nothing to confirm whether the source of the data is being updated
successfully.

This is important because `automysqlbackup`, like `automongodbbackup`, will
silently truncate backups if it runs out of disk space. As demonstrated by:

    root@mysql-backup-1:~# du -csh /var/lib/mysql/signon_production
    439M    /var/lib/mysql/signon_production
    439M    total

    root@mysql-backup-1:~# mount -t tmpfs -o size=8m tmpfs /var/lib/automysqlbackup/
    root@mysql-backup-1:~# df -h /var/lib/automysqlbackup/
    Filesystem      Size  Used Avail Use% Mounted on
    tmpfs           8.0M     0  8.0M   0% /var/lib/automysqlbackup

    root@mysql-backup-1:~# time automysqlbackup /etc/automysqlbackup/automysqlbackup.conf
    Parsed config file "/etc/automysqlbackup/automysqlbackup.conf"

    # Checking for permissions to write to folders:
    base folder /var/lib ... exists ... ok.
    backup folder /var/lib/automysqlbackup ... exists ... writable? yes. Proceeding.

    real    0m37.661s
    user    0m32.522s
    sys     0m2.336s

    root@mysql-backup-1:~# echo $?
    0
    root@mysql-backup-1:~# df -h /var/lib/automysqlbackup/
    Filesystem      Size  Used Avail Use% Mounted on
    tmpfs           8.0M  8.0M  4.0K 100% /var/lib/automysqlbackup

This implementation is a bit horrible because we have to peak inside the
parent script (which `postbackup` is sourced by) to determine the error
code held in `$E`. I've added a fallback which will raise a CRITICAL if `$E`
is unset and should alert us if `automysqlbackup` changes its internals.

We also can't easily lookup what the error code means. I'm going to
subsequently make sure that the email alerts/logs go to somewhere that we
can link to from the Nagios alert.