Reduce Varnish memory usage by another 0.5GB
As a complement to ac06ae00, further reduce Varnish's memory usage by
524MB (1024 - 500).

This is intended as a temporary measure while the low memory issue is
investigated in story #84851852.

Note that Varnish is not the problem here, but it uses the majority of
memory on the cache boxes so it's an easy target to temporarily
alleviate the problem.