Temporarily disable Heka TCP output
Disable Heka's TCP output plugin on all machines.

This is a temporary measure which we hope will prevent Heka from
consuming all of the memory on the boxes it runs on.

Heka will continue to parse logs and push metrics to statsd; in
particular we don't want to lose the HTTP status metrics that we rely on
for alerting.

The central Heka process on the `logging-1.management` box will no
longer receive logs, meaning that Nginx logs will not be visible
centrally in Kibana. Nginx logs will still be retained locally on disk
however and we should be able to catch up with logs when we've resolved
the issues with Heka.

The log retention period on local disk is currently 28 days for Nginx,
which should be enough time to troubleshoot Heka.

See also:
https://www.pivotaltracker.com/story/show/87186816