Send automysqlbackup errors to root/machine.email
Rather than that Zendesk address. I think that queue either no longer exists
or we've all sufficiently filtered it from our inboxes.

We no longer depend on these emails to know when the backup has failed. We
have a Nagios passive check (5761038) to tell us. So these emails are just
to help in debugging those alerts.

Sending these mails to the local `root` account will cause then to be
rewritten to the `machine.email[.plat1]` list in production, which I've
linked to from the Nagios alert.

On staging and preview these mails will stay on the machine because we don't
currently collect machine.email in these environments. I'm not sure there's
much we can do about that right now.