Move Icinga-checking logic to Ruby script
Removing the logic to check for specific statuses
(WARNING, CRITICAL, DOWN and UNREACHABLE) in Nagios across all hosts and
services from a Bash script, and moving it instead in to a Ruby script
we call within the aforementioned Bash script.

Checks for non-200 HTTP status code from Icinga.
Checks for service and host issues; fails with an exit code of 1 if any
found.

Basic JSON parsing checking is included inside hostproblems and
serviceproblems local variables.

Worked on this with @alexmuller, @dancarley, @rjw1 and @mattbostock.

Tested with Taylor Swift references for @alexmuller's enjoyment.