Change Duplicity retention period to fourteen days
We're changing how long we keep Duplicity backups for, to enforce a
retention period of fourteen days rather than one month. This provides us with the
benefit of more regular backups (because we rotate every seven days), and also faster
restoration times because we don't need to use the last monthly backup and every
incremental since until the restoration date.

This does that.

Diff:
https://github.com/alphagov/puppet-duplicity/compare/alphagov:install_pub_key_to_correct_user...use-remove-all-but-n-full-option

:cookie: