replace use_mirror with manage_repo boolean.
use_mirror wasn't set to false anywhere, so remove the option to allow
the upstream elasticsearch module to manage the apt repo.

In it's place, add a manage_repo option to allow disabling the apt
mirror entirely.  This is to allow the support-1 machine to use an
old version of elasticsearch that comes from the deprecated S3 repo.  At
the moment the support-1 machine has the 0.90 elasticsearch mirror
configured, but doesn't install from it.