Don't run Transition Postgres backups on slaves
We shouldn't be running Transition PostgreSQL backups on machines which are
explicity defined as slaves. This is because we cannot guarantee that we can
restore from backups taken from the slave(s), because the master may have
vacuumed data it needs by the time they are taken.

Helpfully, in PostgreSQL, this is easy to work out, because machines hold
their replication state. For instance, 'transition-postgresql-slave' will
always be a slave in the replication set.

To do this, we can set the 'backup' key in Hiera to false, which should stop
backups running from the slave machines. This will then mean that we don't
get alerts about PostgreSQL backups failing on Transition slave machines,
reducing alert fatigue.