Pass heap size setting to init_defaults hash
In #2509 (86667e28 onwards), we changed from our own internal
Elasticsearch module to use the official Puppet Elasticsearch module
from the Puppet Forge.

Our old internal module set Elasticsearch's heap size in the Upstart
script it provided:
https://github.gds/gds/puppet/blob/1af54f7336df53d4c3d2fef25d04fe3cc0f9001b/modules/elasticsearch_old/templates/upstart.conf.erb#L10

But the new official module supports setting the heap size by modifying
a file in `/etc/default`, by passing options to the `init_defaults` hash
which the `elasticsearch::instance` class exposes as a class parameter.

Tested on a Vagrant VM with a 2GB of RAM, hence the heap size is set to
1500MB (75% of total memory). Verified by checking the `-Xms` and
`-Xmx` commandline options set on the java process that runs
Elasticsearch.

  vagrant@logs-elasticsearch-1:~$ ps aux | grep elasticsearch
  110       7329 29.5  7.3 2339840 151372 ?      Sl   15:24   0:03 /usr/lib/jvm/java-7-oracle/bin/java -Xms1500m -Xmx1500m -Xss256k -Djava.awt.headless=true -XX:+UseParNewGC -XX:+UseConcMarkSweepGC -XX:CMSInitiatingOccupancyFraction=75 -XX:+UseCMSInitiatingOccupancyOnly -XX:+HeapDumpOnOutOfMemoryError -Delasticsearch -Des.pidfile=/var/run/elasticsearch-logs-elasticsearch-1.management.development.pid -Des.path.home=/usr/share/elasticsearch -cp :/usr/share/elasticsearch/lib/elasticsearch-0.90.12.jar:/usr/share/elasticsearch/lib/*:/usr/share/elasticsearch/lib/sigar/* -Des.default.config=/etc/elasticsearch/logs-elasticsearch-1.management.development/elasticsearch.yml -Des.default.path.home=/usr/share/elasticsearch -Des.default.path.logs=/var/log/elasticsearch/logs-elasticsearch-1.management.development -Des.default.path.data=/var/lib/elasticsearch-logs-elasticsearch-1.management.development -Des.default.path.work=/tmp/elasticsearch-logs-elasticsearch-1.management.development -Des.default.path.conf=/etc/elasticsearch/logs-elasticsearch-1.management.development org.elasticsearch.bootstrap.ElasticSearch

Note I had to up the Vagrant machine to 2GB of memory before the the
`-Xms` and `-Xmx` values were set correctly; the Vagrant default of
384MB seemingly was not sufficient for Elasticsearch to run.