Enable publishing_api everywhere
content-store currently handles all requests for publishing-api. The
publishing-api app is now feature complete and ready to handle requests.

Before we can make the puppet changes to route requests for publishing-api to
the right app, we need to get it deployed and running everywhere, which is what
this change does.