Document the updating and testing of Mapit
The instructions added in #2141 were not very clear about what happens
after you have created a new SQL dump to be imported into Mapit. It was
assumed that Magic would happen and there was a risk the good
information about testing changes which was in the PR would be lost.