Use puppetlabs/puppetlabs-rabbitmq module
We currently use RabbitMQ from a fork, stored in the Alphagov organisation,
and we have previously referenced this using a branch because we have been
relying on features that haven't been merged upstream.

When we needed to make more changes, we branched from the in-use branch.
Even after the changes we needed were merged upstream, we still continued to
use this fork.

Starting today, we should track from master. The reason we're doing this,
and not pinning in the Puppetfile to a specific version, is that we're
hoping that https://github.com/puppetlabs/puppetlabs-rabbitmq/pull/287 will fix an
issue where RabbitMQ attempts to create a user even if it already exists,
resulting in an error, even though logic exists to not create a user if it
already exists.

This change hasn't yet been rolled in to a point release. This doesn't
restrict the ability to specify versions; they will still be controlled by
the lockfile, and we'll lock around the SHA rather than the branch name.

Specifically, this commit:

- changes the Puppetfile
- changes the Puppetfile.lock to lock to the SHA1 of master as of now
- removes nanliu-staging-0.4* dependencies
- adds nanliu-staging-1.0.3 dependency
- removes tarball of Alphagov fork
- adds tarball of Puppetlabs module