Remove AutoPostgreSQLBackup script
We no longer want to run our automatic PostgreSQL backup script on the slave
machines. We've previously disabled this in Hiera, but since these files
already exist on the machine (and aren't being ensure => absent'd), the
backups still run.

This contains a FIXME to remove this from the node manifest once deployed
throughout the stack.