[#86317146] Don't run PostgreSQL backups on slaves
We shouldn't be running PostgreSQL backups on machines which are
explicitly defined as slaves because if they run at the same time as the
sync job occurs, then the sync job cuts the backups off. This results in
backups that we cannot restore from because they are truncated.

Helpfully, in PostgreSQL, this is easy to work
out because machines hold their replication state. Once a machine is a
slave, forever a slave it shall be unless something occurs to forcibly
change that state.

To do this, we should remove the backup class from the slave manifests.

This was previously, and slightly erroneously, solved using
https://github.gds/gds/puppet/pull/2543 but as @mattbostock pointed out, the
change was actually applied to slaves outside of the Transition PostgreSQL
cluster, so wouldn't have behaved as expected at any rate.

This PR corrects this so that PostgreSQL backups aren't taken from any
slave, and removes the scripts that currently exist on boxes as FIXMEs so we
can go through and remove references to them in Puppet afterward.

It also removes the restriction for backing up MapIt in order to standardise
the PostgreSQL backups across the shop. Whilst we don't need a back-up of
the MapIt data given it's quite small and quite easy to recover from a loss,
having it is useful and means that there are no hidden gotchas out there.