Ignore the reboot needed icinga alert when rebooting.
Currently the unattended reboot script won't reboot a machine unless the
"reboot needed" alert in icinger has been acknowledged.  This is
disingenious as this script is intended to fix that very alert.

This PR therefore makes the check script ignore that specific alert when
checking whether a machine is safe to reboot.