Update test for node classes
This updates the optional spec for all node classes to work with current
versions of everything. This tests all the govuk::node::s_* classes to
ensure that they complie. They run with the real hieradata for the
vagrant environment (except for the s_development class which uses the
development data).

This will provide 2 things:

* It will ensure that all the node classes compile, and should therefore
  catch issues like the one fixed in 3c7e05a.
* It will ensure that the necessary hieradata is populated for the
  vagrant environment, which should ensure that all machines can be
  brought up in vagrant-govuk

This test takes around 12 minutes to run on my machine, and is therefore
left out of the main test run.  It can be run by the separate test:nodes
rake task.