Fix es_dump script shebang for trusty.
/usr/bin/ruby1.9.3 is actually a suymlink to /usr/bin/ruby1.9.1 provided
by the ruby1.9.3 package.  This package actually only provides these
symlinks, and isn't installed by default on a trusty machine.

Instead of relying on these symlinks, switch to use the ruby binary on
the path.  If rbenv is loaded, this will be the rbenv shim, which will
pick up the default rbenv ruby which is the system ruby.  Without rbenv
in the picture, this will pick up the default ruby via the alternatives
mechanism, which is also the correct ruby.