Add new api-elasticsearch cluster running ES 1.4.2.
The main search indices will be moved to this cluster.  This will
achieve 2 things:
* it will move the search cluster into the API vDC which is the more
  correct place for it.
* it will allow us to upgrade rummager to a current version of
  elasticsearch in a controlled manner.

Given that elasticsearch now say that performance should be the same on
Oracle java, and openjdk, this uses openjdk because managing it is a lot
easier.

Note: in anticipation of #2551 being merged, this uses 50% of the RAM
for the ES heap.