Fix elasticsearch sharding config.
The `number_of_replicas` and `number_of_shards` parameters weren't
properly namespaced in the `index` namespace, so were being ignored by
elasticsearch.  This means that all machines will be using the default
values (1 and 5 respectively).

This will only be a problem on the support-1 machines, and the dev VM,
as they are the only places where we override the defaults.

This has been the case since we switched elasticsearch modules in #2509,
and will only have affected indices created since that was deployed.