Cleanup old elasticsearch init script.
With the switch of elasticsearch modules in #2509, the service from the
old module was stopped, but the upstart script wasn't removed.  This
lead to a race condition when machines are rebooted, and in some cases
the old service will start, which causes interesting issues in the
cluster.

This updates the FIXME to ensure the old service is stopped, and then
remove the upstart script.