[publishing-api] configure backend(-lb) to handle publishing-api
This is the first step in switching apps over to use the publishing-api, rather
than content-store (which currently has an alias for publishing-api).

This change should mean that if a request is routed to the backend-lb servers,
then it is handled and routed to the backend servers instead of being denied.

Once this change has been rolled out, we can change the hosts files so that
requests for publishing-api get routed to backend-lb.

Once that change is rolled out, we can tidy up the now-redundant publishing-api
alias on api-lb and content-store servers.