[#87001376] Restart Rsyslog if logs stop arriving
Having monitoring for CDN logs isn't much use if they break in the evening
or at weekends. So far, in quite a few instances where they have broken, this
has been the case. Since restarting `rsyslogd` is usually the fix for the problem,
we should automatically check when log acceptance is broken and restart
`rsyslogd`. This will hopefully minimise the amount of lost data.

Restarting `rsyslogd` will stop other logs from being received from Fastly,
however the Core team are only concerned about the logs we get from Bouncer
so this is a price worth paying.

This commit adds a shell script to query NRPE to see if the log file is
older than five minutes. If it is, then we log to syslog for visibility then
restart the rsyslog service. This shell script runs on cron every five
minutes.

We also change the Icinga check limits so that we warn after 30mins and
critical after 1hr, meaning that we should be able to catch when things
really break much more quickly. The likelihood of things really breaking is
somewhat reduced given that we are now automating some of the fixing in this
commit.

This story is closely linked to #88888468 in the WebOps backlog.