Amend automongodbbackup script log verbosity level
Per https://www.pivotaltracker.com/story/show/88131008, I want to increase
the log verbosity of this script so we can try to see at what point it
errors and why.

Mail is sent to a Google Group, so hopefully we can leave this for a few
days and corroborate any issues between logs, unless something very obvious
is noticable beforehand.