content-store no longer responds to publishing-api
content-store was aliased to handle requests for publishing-api. Now that we
are routing requests for publishing-api to publishing-api (bae64dcdfcbf8a61e622c6bc4933d6dcc6222a37),
we can remove support from content-store's nginx configuration and the API load
balancers for handling such requests.

This change includes making the change in development as they are so closely
intertwined.